requirements.txt
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
